create stringUtils class
